+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add a CSV export for Academies due to transfer in a time period (this work is
   not publicly viewable)
 - Add reduced support service banner to the service for the christmas period
+- Build the export journey for Academies due to transfer in a time period
 
 ### Changed